# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AssetFlow - Enterprise Asset & Resource Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all assets (excludes soft-deleted)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [available, allocated, reserved, under_maintenance, lost, retired, disposed]
        - name: condition
          in: query
          schema:
            type: string
            enum: [excellent, good, fair, poor]
        - name: bookable
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create an asset (tag auto-generated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get an asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Soft-delete an asset (sets status to disposed, never permanently removes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Soft-deleted asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tag:
          type: string
          description: Auto-generated tag (e.g. AF-000001)
        serialNumber:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        employeeId:
          type: ["integer", "null"]
        employeeName:
          type: ["string", "null"]
        purchaseDate:
          type: ["string", "null"]
          format: date
        purchaseCost:
          type: ["number", "null"]
        currentValue:
          type: ["number", "null"]
        vendor:
          type: ["string", "null"]
        condition:
          type: string
          enum: [excellent, good, fair, poor]
        status:
          type: string
          enum: [available, allocated, reserved, under_maintenance, lost, retired, disposed]
        bookable:
          type: boolean
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        currentAgeMonths:
          type: ["integer", "null"]
        warrantyRemainingMonths:
          type: ["integer", "null"]
        deletedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - tag
        - condition
        - status
        - bookable
        - createdAt
    AssetInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        serialNumber:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        employeeId:
          type: ["integer", "null"]
        purchaseDate:
          type: ["string", "null"]
          format: date
        purchaseCost:
          type: ["number", "null"]
        currentValue:
          type: ["number", "null"]
        vendor:
          type: ["string", "null"]
        condition:
          type: string
          enum: [excellent, good, fair, poor]
          default: good
        status:
          type: string
          enum: [available, allocated, reserved, under_maintenance, lost, retired, disposed]
          default: available
        bookable:
          type: boolean
          default: false
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - name
    AssetUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        serialNumber:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        employeeId:
          type: ["integer", "null"]
        purchaseDate:
          type: ["string", "null"]
          format: date
        purchaseCost:
          type: ["number", "null"]
        currentValue:
          type: ["number", "null"]
        vendor:
          type: ["string", "null"]
        condition:
          type: string
          enum: [excellent, good, fair, poor]
        status:
          type: string
          enum: [available, allocated, reserved, under_maintenance, lost, retired, disposed]
        bookable:
          type: boolean
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...